backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
dnspython==2.8.0
dotenv==0.9.9
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.116.1
greenlet==3.2.4
idna==3.10
passlib==1.7.4
psycopg2==2.9.10
pyasn1==0.6.1
pydantic==2.11.9
pydantic-core==2.33.2
python-dotenv==1.1.1
python-jose==3.5.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.43
sqlmodel==0.0.24
starlette==0.47.3
typing-extensions==4.15.0
typing-inspection==0.4.1
